NewRelic APM Transaction
Abstraction for NewRelic transactions. Our intent is to provide a
way to register transactions on NewRelic with portion of php code
or long running php process. You can specify a new Transaction and
call a method it will start just after it is called and will end
just before it return.