Allow classes to be partially monitored
There are some bad designs that use multiple calls to a consumer class
trying to ask the class what to do instead of tell it to perform the job
and this lead to call __call multiple times generating one transaction
per method called.

Now you can specify the name of the method you want to spy. For backward
compatibility the method name is empty maitaining same behavior as before
which is transact every method call.